fix docker tag "latest" to point to newest release (#64)
Previously, the docker tag "latest" on quay.io was moved each time a new
commit was pushed to the "master" branch on GitHub. This didn't match
the usual convention that "latest" points to the most recent official
release.

This commit changes the meaning of docker tags on quay.io to the
following:

- "latest" - will point to the newest "vX.Y.Z" release (according to
  SemVer rules)
- "dev" - will point to an image built from the most recently changed
  commit on GitHub

Note: the change to the meaning of the "latest" tag will only happen
when the next git tag is pushed to the GitHub repository.