dynamically pin base image versions in Dockerfiles (#64)
In template processors Dockerfiles, make sure that appropriate versions
of base eunomia images are used. Previously, always the "latest"
versions of eunomia base images were used, which was incorrect from
reproducibility point of view.

Also, build the base repository URLs for the base images dynamically.
This makes it easier to run the build-test-images Makefile target in a
non-KOHLS Quay.io namespace, for testing.