scripts/e2e-test.sh: fix to use correct images in e2e test
Previously, the template-processors images used in e2e-test.sh were
downloaded from Quay.io. As a result, the e2e-test.sh script did *not*
actually test the whole current commit "end to end" - because the code
of template processors from the current commit was ignored.

The idea behind this commit is to pre-populate the docker repository
used by the minikube cluster, such that it doesn't try to download
images from Quay.io. This was implemented through the following changes:
 - scripts/e2e-test.sh - modified to build images and store them in
   minikube's docker repository
 - *.yaml - all "imagePullPolicy: Always" are changed to "IfNotPresent",
   to make it possible for minikube to not download images from Quay.io
 - Makefile - the e2e-test-images target was extended to build with a fake
   "v999.0.0" version tag, to simulate a regular release, and force
   "latest" docker tags to be generated by scripts/build-images.sh
 - .travis.yml - the "make e2e-test-images" step is removed, as it is
   now done in scripts/build-images.sh

This is unfortunately a tricky and subtle change, but I don't have a
good idea how to make it simpler and more robust as of yet.