.travis.yml: pin minikube version 1.3.1
After minikube v1.4.0 was released, builds started to fail on `minikube
update-context`. Pin minikube to the previously working v1.3.1 as a
workaround. The old script was based on a tutorial on travis-ci.com
blog, so it is either a regression, or v1.4.0 needs some new approach.

Example build succeeding with v1.3.1:
  https://travis-ci.com/KohlsTechnology/eunomia/builds/128181911#L637
Example build failing with v1.4.0:
  https://travis-ci.com/KohlsTechnology/eunomia/builds/128748234#L637