deploy/helm/prereqs/templates: fix ClusterRole to allow namespaces listing (#115)
The manageDeletion function in
pkg/controller/gitopsconfig/gitopsconfig_controller.go, which implements
eunomia-operator's finalizer, needs read access to cluster's namespaces.
This access was missing, which resulted in the finalizer hanging.

Fixes #115.

NOTE: Why this lack of permissions resulted in eunomia being actually
*stuck* (hanging) in the finalizer, is currently really weird to me.
Specifically, with the missing permissions, the kubernetes API client
library was repeatedly trying to redo the operation, instead of just
returning an error after a first failure. What is even more weird to me,
is that even though I tried adding a timeout on the Context objects used
by the client in a couple places, I still wasn't able to force it to
break this retry loop.

I am considering drilling down on this behavior to try and understand
the source of it, and thus to hopefully protect us against similar
problems in future, but it has proven to be surprisingly elusive as of
yet. Thus, I decided to put that investigation on hold, and get back to
submitting just a fix for now.