Update TravisCI helm publishing related scripts
This commit updates scripts/build scripts related to publishing of helm charts after creation of tag in repo. In scripts/build/build-chart-repo.sh it removes line templating values.yaml file as we are not templating it in current chart. In scripts/build/push-to-pages.sh it sets git to explicitly use github_deploy_key while connecting to GitHub.