template-processors: set -eu[x]o pipefail
Add bash flags protecting against common errors in all scripts (in
template-processors/ and in scripts/ too). Also add `set -x` where
possible, to make potentially debugging them easier in future.

As part of adding the flags, a place was found in discoverEnvironment.sh
where `kubectl describe` command was failing. The place was marked with
FIXME note and a separate issue will be opened for fixing it.

Closes #74.