.github: redesign ISSUE_TEMPLATE.md based on one in golang repository
The order of questions in the old template had some issues.
Specifically, a reporter trying to answer them in order, if they wanted
to provide some basic context, would have to repeat a lot of information
between the answers. For example, to answer "What happens?" question,
a reporter might want to start by describing what they did. Which
duplicates the work they'd then have to do in "Steps to reproduce"
section, without clear guidance as to how to structure the answer or how
much detail to provide.

This commit changes the template, based on ISSUE_TEMPLATE.md from
https://github.com/golang/go repository, adapted to our case. The
questions in this one are laid in such an order, that answers to the
earlier ones provide context on which the answers further down can
easily build, without repeating information.