test/e2e: make timeouts be consts in main_test.go
The timeouts used in various tests don't need to be variables, they can
be consts, as they're not expected to be changed by particular tests.

Also, given that they're used by more than just one tests, move them to
a main file of the package, to communicate better that they're a shared
part of the package code.