test/e2e: fix and refactor GetPod and its callers
The WaitForPodWithImage function had a bug, where it was written as if
GetPod could return an error, though it could not. Comparing it with a
similar WaitForPod func, it looked useful to allow returning an error
from GetPod, to help in debugging if needed.

Also, the WaitFor... funcs are refactored to make them easier to
understand.