template-processors/base: delete resources removed from git
This commit adds a feature of auto-detecting which resources were
removed from input yaml files inbetween eunomia runs. From now on, they
will be automatically deleted from the cluster. Without this, it was
required of user to manually track resources they wanted to delete, and
explicitly order eunomia to do so; otherwise they'd be kept untouched in
the cluster after they were removed from the git repo.

The high-level overview of the implementation, which is mostly done in
resourceManager.sh:
 1. We modify all the input .yaml files, adding two extra labels to each
    Kubernetes resource they contain:
      - "owner": contains unique id of the GitOpsConfig CR
      - "applied": contains a timestamp of when the resource was last
        applied (with `kubectl apply`) by the resourceManager script
 2. After applying the resources (with `kubectl apply`), we search for
    all the resources in the cluster for which:
      - "owner" corresponds to the currently applied GitOpsConfig CR
      - but "applied" *differs* from the currently applied timestamp
        (which means they were not present in the current run of
        resourceManager)
    and delete them from the cluster (`kubectl delete`).

Fixes #24.