chore: run e2e tests in CI from correct branch for PRs
For e2e-test.sh script, it is necessary to set EUNOMIA_REF to a correct
commit in the git repo containing templates used by template processors.
For PRs changing those files, this needs to be set to the commit of the
PR. The $TRAVIS_PULL_REQUEST_SHA variables we'll be using is documented
as getting set to a commit SHA1 for PRs, and left empty for non-PR
builds.