Set Docker context to be operator-sdk compliant
Set Docker context in building eunomia-operator image to eunomia project
root directory, and pass Dockerfile location explicitly with the -f
flag.

Docker context in eunomia-operator was set to eunomia/build which was
causing "operator-sdk build" command to not work.

Now, with the context set properly, eunomia-operator Docker image can be
built either with the make utility, or "operator-sdk build" command.

The issue this commit is resolving is verbosely explained here:
https://github.com/KohlsTechnology/eunomia/issues/174