Resort to default job templates if configmaps miss
Copy the default job templates in Dockerfile. Use them in case of
configmaps missing. Log a proper message.
Also remove any volumes created from the configmaps both in config files
and tests.

Thanks to this configmaps are not obligatory for deploying
eunomia-operator. It can be now seamlessly deployed. This paves the way
for pushing eunomia to operatorhub:
https://github.com/KohlsTechnology/eunomia/pull/86