chore: create `make generate` target
In Makefile, add new target `generate`, which will run code-generation
subcommands of operator-sdk 0.8.1 (same version as is currently defined
in .travis.yml). Thus, running `make generate` will ensure that files
controlled by operator-sdk are up to date. Previously, it was not easy
to find out how to regenerate those files when needed.

The 0.8.1 version of operator-sdk we are currently using is fairly old,
so it's not easily available in default OS package managers for
contributors to use; therefore, docker is used to easily ensure the
command generates the same code for everybody.