Fix local and Travis e2e tests
Set image pull policy to Never. Thanks to this already present image
will not be pulled from external registry which was the case so far. If,
for some reason, the image is not present locally (and it should always
be), an error will be thrown.

Set "GOOS=linux" before "make e2e-test-images". Thanks to this when the
tests are run locally on system other than Linux, no error will be
thrown in minikube VM which has Linux architecutre. Before this commit
it was not a problem because the images were always pulled from external
registry. And those images are built for Linux architecture.