chore: rename `make fmt` target to `check-fmt`
The `fmt` name of the target was misleading, as the usual convention is
to have `make fmt` change the code in the repository to make it conform
to gofmt rules. (I noticed this expectation also in real life from a
contributor to eunomia.) The old `fmt` target was not modifying the
repository, but instead checking if the repo conforms to gofmt rules.
The new name of the target reflects this intention.