chore: add `make test-dirty` target and run it in CI
Add a make target which verifies that `make generate` was called on
contributed code. This is a test that may modify the code, so it is not
included in normal `make test`, to not override people's code when
running tests locally. This test is however enabled as part of the CI.

Additionally, in `test-dirty`, verify that the code built by CI is done
from a clean repository, so that no unexpected changes sneak into the
builds.