pkg/controller/gitopsconfig: emit k8s Events on job completion (#110)
This commit makes eunomia emit Kubernetes events when Jobs it controls
end. This is expected to be useful for tracking when a particular GitOps
change got applied, and whether it applied successfully.

The implementation is based on watching Job resource change events
reported by the Kubernetes client-go API. This commit does not reuse the
mechanisms established in jobmonitor.go. The reasons for this are
multi-fold:
  - This commit is an exploration of the possibility of using Kubernetes
    watching mechanisms (Informers), as a next step after the
    jobmonitor.go code. I believe this approach has a chance to be more
    robust and in-line with the features exposed by the client-go API. I
    hope jobmonitor could be refactored to this same mechanism in future
    as well.
  - Comparing the two approaches, it seems that the one used in this
    commit detects more job completion events than the one used in
    jobmonitor.go.

This commit is a partial fix to #110. However, it does not yet implement
all its requested features:
  - no Git SHA1 in the Message - this was postponed for the time being,
    to reduce the scope of the current task
  - InvolvedObjects field is not used - the Events API as exposed in
    client-go doesn't seem to provide means of setting this field;
    instead, this commit places extra data (Job name) in the Annotations
    field.