pkg/controller/gitopsconfig: update Status based on informers (#217)
This commit changes the mechanism for updating the Status field of
GitOpsConfig from cron (pull-based) to informers (push-based).

The implementation in this commit is unfortunately not ideal. I
encountered many challenges when developing it, and in the end took
compromises in some aspects. Specifically:

1. Status is now calculated based on single Job's data received from
   Informer. As such, this makes Status calculations "edge-based",
   instead of "level-based", with the latter being the recommended
   approach for Kubernetes apps. In case Informer misses some changes
   (for example, potentially in case of lost network connection), we may
   miss recalculating some GitOpsConfig Status transition.

   I tried changing this approach to "level-based". In OnUpdate, I tried
   List-ing all Jobs belonging to GitOpsConfig, and calculating Status
   based on the newest one among them. However, to my surprise, List was
   giving me *stale data* from Kubernetes. I considered adding an extra
   cron-like task, that would periodically recalculate Status for all
   GitOpsConfig CRs; however, it's not yet clear to me how to implement
   it in a stateless and robust way in eunomia.

   In the end, I reverted back to the "edge-based" approach as shown in
   this commit, finding it the simplest implementation for now, while
   seemingly (to my surprise) giving robust results. I am considering a
   theory, that maybe the Informers infrastructure is robust enough that
   it takes some complexity from its users.

2. Setting GitOpsConfig Status can currently fail in case the CR was
   modified between the Get() and Status().Update() calls. Current
   implementation doesn't do any retries in such case, thus making it
   possible for Status to become stale and incorrect.

   However, it's not clear to me how to implement retries correctly.
   I expect retries shouldn't block the OnUpdate function with
   Sleep-like mechanism, as that would risk stalling that part of
   eunomia. Also, I am not sure how to do it in a stateless way, such
   that it would retry also in case eunomia is restarted in meantime.

   Again, I would consider trying to use a cron-like mechanism here,
   possibly as complementary to "fast" edge-based updates. However, I
   see challenges in how to implement such cron well in a stateless and
   scaling-safe way, so I decided to not do it yet, and bring it up to
   discussion.

As a potential workaround, if nobody comes up with a better idea, I
would like to propose adding a background thread in eunomia, repeatedly
picking a random GitOpsConfig from the cluster, and recalculating its
Status. That could be a method to achieve eventual consistency, though
in a non-deterministic way.

Additionally, through the course of implementing this feature, I came to
believe it would be good to simplify tracking the relation between
GitOpsConfig, and Jobs it spawns, by marking the Jobs with a new label,
with GitOpsConfig name as its value. This would simplify the "owner
detection" code in numerous places in eunomia, currently done in
numerous different ways, and also make it simpler and more future-proof.

I wanted to make it part of the work done in this commit, but I decided
to postpone it for now, as implementing this feature as is was already
taking me too long.