scripts/e2e-test.sh: add --verbose flag
I got a situation on Travis-CI where a build failed because of no output
from `operator-sdk test local` for over 10 minutes:

https://travis-ci.com/KohlsTechnology/eunomia/builds/144500591

As we increase the number of test, their running time also increases.
Adding the --verbose flag aims to get some progress output printed on
Travis CI console, so that the pipeline is not stopped prematurely.

See also:
- https://docs.travis-ci.com/user/common-build-problems/#build-times-out-because-no-output-was-received