pkg/controller/gitopsconfig: don't panic in statusUpdater for unstarted jobs (#217)
The statusUpdater.OnUpdate method was causing a nil panic when it
received a Job with empty StartTime. Fix this issue by ignoring such a
Job.

The decision to ignore (instead of trying to set Status with empty
StartTime) is because we need StartTime to resolve ordering of Status
updates, to avoid updating Status with data from some too old Job.
Without StartTime, we can't do this.