build/job-templates: add "jobOwner" label on jobs
Add a label on jobs, referencing the GitOpsConfig that created them.
This is intended to make it much easier to find Jobs & CronJobs owned by
GitOpsConfig, as well as GitOpsConfig that is owner of a Job/CronJob.
Previously, finding parent/children in this relation was done by ad-hoc
traversing of the objects graph in kubernetes. With this change, it
should be possible to do it with a single List query to kubernetes with
a label filter.

This will be used in a later commit fixing issue #216. Additionally,
it is expected that this can simplify code in a few more places in
GitOpsConfig controller, including status_updater.go and jobwatch.go.