pkg/controller/gitopsconfig: use label to find CR's active Jobs
Modify the way active Jobs are searched for a GitOpsConfig CR. Use a
search by label, instead of iterating over jobs and checking ownership.
This is a conceptually simpler and more universal approach (can be used
in many places in very similar way, instead of custom complex per-case
helper functions).