scripts/run-eunomia-locally: add the script
Add script to make it easier to develop eunomia locally. Previously the
user had to start minikube and set the env variables himself. Now this
is done for him in this script.

Using of `operator-sdk up local` command enforces the use of a
Kubernetes cluster so there is no option to use OpenShift.