deploy: include eunomia-cluster-list clusterrole
When Eunomia jobs are trying to determine what resources it needs to
delete, it retrieves all kubernetes resource kinds, and uses that list
of kinds to get all resources that meet the "ownership" and the "not
applied in this job" filters.

If the service account the job is running as does not have a ClusterRole
that provides access to list cluster scoped resources, the job will log
`Forbidden` errors for every one of the cluster scoped resource types
which can be confusing to users.

This ClusterRole gives the job quite a bit of access, so in a future
pull request it would make sense to only search in kinds that the
service account has access to manage.

Fixes #240