scripts: add basic test for ocp-template processor
Add a basic test is to ensure template processor can work on minishift.
The script is currently not yet part of CI. In its current state, the
script is very rudimentary. I only wanted to get it to a point, where on
my machine it manages to successfully create a helloworld Pod in
minishift. I assume the script will be improved in future until it gets
to a state where it can be merged into the CI pipeline.