pkg/controller/gitopsconfig: fix go test flags support (#264)
Previously, `go test` flags (e.g. `-v`) were not honored by the tests in
pkg/controller/gitopsconfig because of complex tests initialization.
This commit removes some initialization steps in order to re-enable `go
test` flags support.

Also, the initialization is moved from TestMain to init func. This is
not a crucial change, but it allows us to remove the TestMain function
and rely on the default one, while making the contents of the resulting
init func slightly simpler.

Related to #264