scripts/operatorhub-deploy: Add the script
Add the script operatorhub-deploy.sh for the user to be able to deploy
new Eunomia version to OperatorHub without effort. All files needed by
operator-sdk are being automatically generated or downloaded and in the
end cleaned up. The result of running this script is a PR in the
https://github.com/operator-framework/community-operators repository.

Also add a simple howto in .github/RELEASE.md.