gitopsconfig: make sure bad img blocked job
Double-check if the existing job is blocked because of bad image. Add
checking if there is only one active pod of a job, which occurs when
nonexistent or wrong image blocks the job. Thanks to it only such case
will fall under the if code block.
Before, if there was a "normal" fail of a job (i.e. code inside
container exited 1), and there were a few failed pods, this would fall
under the if block, and jobContainerStatus function would return an
error because the job had a few pods.