pkg/controller: add shorthand alias for types.NamespacedName
The types.NamespacedName struct from kubernetes API is used very often
in code. It's quite cumbersome to write and noisy when reading. This
commit introduces a type alias for the struct, named util.NN. The new
alias is intended to make the code easier to write and read (less
"visual stuttering"), while still keeping the information that was
present before available to reader. The new alias is intended to be used
in a way resembling "named arguments" to functions in other programming
languages. For example, new code:

    someFunc(..., util.NN{Name: "foobar", Namespace: "barfoo"}, ...)

vs. code it's intended to replace:

    someFunc(..., types.NamespacedName{Name: "foobar", Namespace: "barfoo"}, ...)

Notice that in the old code, words "Namespace" and "Name" are repeated
twice. In the new code, they're still present, so the information is
retained, while they're present only once, without noisy repetition
("stuttering").