template-processors/base: use fully qualified kind names in kubectl (#273)
Before, a list of resource kinds passed to `kubectl delete` was built
based on data from "kind" column of `kubectl get`. However, this showed
up to be bad, as the data was not fully & uniquely identifying for
resource kinds. As a result, reusing them later in `kubectl delete`
resulted in different kinds being targeted than expected.

This commit changes the approach to use jsonpath formatting of `kubectl
get` to extract raw "kind" and "apiVersion" data. This data allows us to
build a unique, fully qualified resource name. Unfortunately, it
requires some extra processing with `awk`, as the format expected by
`kubectl get` differs somewhat from what's present in
"kind"+"apiVersion".

Unfortunately, I didn't manage to find a way to write an automated e2e
test for this fix.

Fixes #273.

Alternatives
------------

The approach chosen in this commit seems to work, but it's not 100%
clear to what extent it will be robust. It relies on "kind" and
"apiVersion" fields returned by resource controllers in response to
"list" requests, however those responses seem to not be 100% guaranteed
to be correct. One alternative approach could then be to run separate
`kubectl get` for each known kind, and if 1+ results were returned,
remember the kind used in the request (vs. response as is done
currently). However, when I tested this approach, it had a *huge* impact
in slowing down script completion time. Because of that, I chose the
simpler and faster approach for now. I believe we should see how it
works in practice, and decide on the slower approach only in case the
current approach shows up to be not enough.