chore: restore go.mod after `operator-sdk generate` (#246)
Currently, because we're using old operator-sdk for `operator-sdk
generate` in Makefile, we have to use Go 1.12 (to properly put code it
depends on in 'vendor' directory). Calling Go 1.12 from a Go 1.13
project results in big changes in go.mod. This commit adds commands for
restoring go.mod from before the `operator-sdk generate` call.

Related to #246.