chore: upgrade operator-sdk to v0.17.1
A dependency for a dependency for Operator-SDK v0.12.1 (bitbucket.org/ww/goautoneg)
has been moved, so we have to update the operator-sdk version so we can build Eunomia.
Since it is not a dependency directly in our code, we are not able to
simply replace the dependency in the go.mod.

Since CustomResourceDefinition has been promoted from v1alpha1 to v1 in
Kubernetes version 1.16, Eunomia only supported on Kubernetes version
1.16 and on. The test matrix has been update to reflect this.

In addition, version 1.6 of minikube does not work with Kubernetes version
1.17 and on, so the TravisCI is being updated to use the latest version.

Fixes #373