pkg/handler: improve clarity in webhook handler
Logging improvements provide better clarity as to why a GitHub push
event did not trigger a reconsiliation job for each GitOpsConfigs.

A new check was also added to see if any GitOpsConfig match the webhook
event payload. If no GitOpsConfigs match we log there were no
GitOpsConfigs found and return. This prevents the log message stating
that we are reconciling a GitOpsConfig with an empty namespace and
empty GitOpsConfig name mentioned in issue #270:

{"level":"info","ts":#,"logger":"gitopsconfig-controller","msg":"Reconciling GitOpsConfig","Request.Namespace":"","Request.Name":""}