chore: include go mod vendor directory
This will allow us to rebuild old versions of Eunomia, even if
dependencies are moved.

The script `scripts/update-crds.sh` is replacing the docker image that
was used to simplify the process as the environment required to generate
resources via operator-sdk and openapi-gen are minimal.

The make target `test-dirty` has also been updated to ensure that the
vendor directory is up to date.

Fixes #373

check vendor directory up to date